修订代码；为 quic的 0-rtt功能做铺垫.
