+++ examples/trojan_grpc_smux.client.toml
@@ -9,7 +9,6 @@ protocol = "trojans"
 uuid = "a684455c-b14f-11ea-bf0d-42010aaa0003"
 host = "127.0.0.1"
 port = 4434
-version = 0
 insecure = true
 utls = true
 advancedLayer = "grpc"
